+++ src/probnum/linalg/solvers/belief_updates/solution_based/_solution_based_proj_rhs_belief_update.py
@@ -5,7 +5,7 @@ from probnum import randvars
 from probnum.linalg.solvers.beliefs import LinearSystemBelief
 from probnum.typing import FloatArgType
 
-from ._linear_solver_belief_update import LinearSolverBeliefUpdate
+from .._linear_solver_belief_update import LinearSolverBeliefUpdate
 
 
 class SolutionBasedProjectedRHSBeliefUpdate(LinearSolverBeliefUpdate):
