+++ tests/test_linalg/test_solvers/cases/belief_updates.py
@@ -2,17 +2,17 @@
 linear system."""
 from pytest_cases import parametrize
 
-from probnum.linalg.solvers import belief_updates
+from probnum.linalg.solvers.belief_updates import matrix_based, solution_based
 
 
 @parametrize(noise_var=[0.0, 0.001, 1.0])
 def case_solution_based_projected_rhs_belief_update(noise_var: float):
-    return belief_updates.SolutionBasedProjectedRHSBeliefUpdate(noise_var=noise_var)
+    return solution_based.SolutionBasedProjectedRHSBeliefUpdate(noise_var=noise_var)
 
 
 def case_matrix_based_linear_belief_update():
-    return belief_updates.MatrixBasedLinearBeliefUpdate()
+    return matrix_based.MatrixBasedLinearBeliefUpdate()
 
 
 def case_symmetric_matrix_based_linear_belief_update():
-    return belief_updates.SymmetricMatrixBasedLinearBeliefUpdate()
+    return matrix_based.SymmetricMatrixBasedLinearBeliefUpdate()
