+++ tests/test_linalg/test_solvers/test_belief_updates/test_matrix_based/test_matrix_based_linear_belief_update.py
@@ -1,4 +1,4 @@
-"""Tests for the symmetric matrix-based belief update for linear information."""
+"""Tests for the matrix-based belief update for linear information."""
 
 import pathlib
 
@@ -9,23 +9,21 @@ from pytest_cases import parametrize_with_cases
 from probnum import linops, randvars
 from probnum.linalg.solvers import LinearSolverState, belief_updates, beliefs
 
-case_modules = (pathlib.Path(__file__).parent / "cases").stem
+case_modules = (pathlib.Path(__file__).parent.parent / "cases").stem
 cases_belief_updates = case_modules + ".belief_updates"
 cases_states = case_modules + ".states"
 
 
 @parametrize_with_cases(
-    "belief_update",
-    cases=cases_belief_updates,
-    glob="symmetric_matrix_based_linear*",
+    "belief_update", cases=cases_belief_updates, glob="matrix_based_linear*"
 )
 @parametrize_with_cases(
     "state",
     cases=cases_states,
-    has_tag=["has_action", "has_observation", "symmetric_matrix_based"],
+    has_tag=["has_action", "has_observation", "matrix_based"],
 )
 def test_returns_linear_system_belief(
-    belief_update: belief_updates.SymmetricMatrixBasedLinearBeliefUpdate,
+    belief_update: belief_updates.matrix_based.MatrixBasedLinearBeliefUpdate,
     state: LinearSolverState,
 ):
     belief = belief_update(solver_state=state)
@@ -33,15 +31,15 @@ def test_returns_linear_system_belief(
 
 
 @parametrize_with_cases(
-    "belief_update", cases=cases_belief_updates, glob="symmetric_matrix_based_linear*"
+    "belief_update", cases=cases_belief_updates, glob="matrix_based_linear*"
 )
 @parametrize_with_cases(
     "state",
     cases=cases_states,
-    has_tag=["has_action", "has_observation", "matrix_based"],
+    has_tag=["has_action", "has_observation", "symmetric_matrix_based"],
 )
-def test_raises_error_for_non_symmetric_Kronecker_structured_covariances(
-    belief_update: belief_updates.SymmetricMatrixBasedLinearBeliefUpdate,
+def test_raises_error_for_non_Kronecker_structured_covariances(
+    belief_update: belief_updates.matrix_based.MatrixBasedLinearBeliefUpdate,
     state: LinearSolverState,
 ):
     with pytest.raises(ValueError):
@@ -49,15 +47,15 @@ def test_raises_error_for_non_symmetric_Kronecker_structured_covariances(
 
 
 @parametrize_with_cases(
-    "belief_update", cases=cases_belief_updates, glob="symmetric_matrix_based_linear*"
+    "belief_update", cases=cases_belief_updates, glob="matrix_based_linear*"
 )
 @parametrize_with_cases(
     "state",
     cases=cases_states,
-    has_tag=["has_action", "has_observation", "symmetric_matrix_based"],
+    has_tag=["has_action", "has_observation", "matrix_based"],
 )
 def test_against_naive_implementation(
-    belief_update: belief_updates.MatrixBasedLinearBeliefUpdate,
+    belief_update: belief_updates.matrix_based.MatrixBasedLinearBeliefUpdate,
     state: LinearSolverState,
 ):
     """Compare the updated belief to a naive implementation."""
@@ -67,19 +65,15 @@ def test_against_naive_implementation(
     ):
         pred = matrix.mean @ action
         resid = observ - pred
-        covfactor_Ms = matrix.cov.A @ action
+        covfactor_Ms = matrix.cov.B @ action
         gram = action.T @ covfactor_Ms
         gram_pinv = 1.0 / gram if gram > 0.0 else 0.0
         gain = covfactor_Ms * gram_pinv
-        covfactor_update = gain @ covfactor_Ms.T
-        resid_gain = np.outer(resid, gain)
+        covfactor_update = np.outer(gain, covfactor_Ms)
 
         return randvars.Normal(
-            mean=matrix.mean
-            + resid_gain
-            + resid_gain.T
-            - np.outer(gain, action.T @ resid_gain),
-            cov=linops.SymmetricKronecker(A=matrix.cov.A - covfactor_update),
+            mean=matrix.mean + np.outer(resid, gain),
+            cov=linops.Kronecker(A=matrix.cov.A, B=matrix.cov.B - covfactor_update),
         )
 
     updated_belief = belief_update(solver_state=state)
